Replace use of "stakeholder"
This term is increasingly being reconsidered in light of it's negative
historical usage. See, e.g.,

* https://www.fasttrackimpact.com/post/why-we-shouldn-t-banish-the-word-stakeholder?postId=1a6b9631-56a5-416e-a796-86d6dfb30d87
* https://researchimpact.ca/featured/switching-from-stakeholder/

This replaces the usage for all but the Ethics material, where the word
is actively used throughout the reading. An issue has been opened to
address if more actively there:

https://github.com/weecology/forecasting-course/issues/53

Thanks to the Ecological Forecasting Initiative DEI group, for
their discussion that precipitated this change.